# Split TreeWalker to Visitor and getChildNodes