… packages/pyright-internal/src/analyzer/parseTreeWalker.ts