This is the first step to refactor checker to use 1 walker but multiple visitors which do its own checking inside of the checker.

PR with perf test (https://github.com/microsoft/pyrx/pull/2895)

and the result is pasted below.

`original walker` is what we have now 
`heap based walker (with 1 switch)` is a heap based (no recursion) [walker](https://github.com/microsoft/pyrx/blob/09287d8d55320e9b33215ba280696480b3e3130f/packages/pyright/packages/pyright-internal/src/analyzer/parseTreeWalker4.ts#L100) of the `original walker`
 `refactored walker (with 2 switch)`  is what I have in this PR
`heap based walker (with 2 switch)` is a heap based [walker](https://github.com/microsoft/pyrx/pull/2895/files#diff-a438e8c98f3ef4ee1f33c7a54814400af89b6bf6a60e26d4abe95358c7be8044) for `refactored walker (with 2 switch)`

I believe all perf difference comes from the fact that `original walker` had only 1 big switch to dispatch node based on node kind, but  `refactored walker (with 2 switch)`  has 2 of those (visit and getChildNodes)

converting recursion based to heap base didn't improve perf probably because it now has a real stack in heap to hold nodes to visit.

and as you said, the `original walker` is the fastest one, about 10% to 20% faster than other ones.

but that being said, it is like avg 0.02-0.04ms differences in elapsed time per tree or 0.00001-0.00002ms difference per node. which won't be noticeable. especially if walker does actual work rather than just purely iterating through all nodes.

```
  ~#@❯  npm run perftest                                                            

> pylance-internal@9999.0.0-dev perftest
> ts-node --project ./tsconfig.withBaseUrl.json -r tsconfig-paths/register ./src/walkerPerf.ts

scan files in D:\pytest\requestTest\.env\Lib\site-packages
found 4998 files


run 1 *******
run gc to clean memory

total file size 69548977bytes, total parse time 15.14sec for 4998 trees (3.03ms per tree)
original walker
total walking time 1.21sec for 4998 trees and 8770308 nodes. (0.24ms per tree) (0.000138ms per node)
heap based walker (with 1 switch)
total walking time 1.09sec for 4998 trees and 8770308 nodes. (0.22ms per tree) (0.000124ms per node)
refactored walker (with 2 switch)
total walking time 1.19sec for 4998 trees and 8770308 nodes. (0.24ms per tree) (0.000135ms per node)
heap based walker (with 2 switch)
total walking time 1.11sec for 4998 trees and 8770308 nodes. (0.22ms per tree) (0.000126ms per node)



run 2 *******
run gc to clean memory

total file size 69548977bytes, total parse time 15.63sec for 4998 trees (3.13ms per tree)
original walker
total walking time 0.93sec for 4998 trees and 8770308 nodes. (0.19ms per tree) (0.000105ms per node)
heap based walker (with 1 switch)
total walking time 1sec for 4998 trees and 8770308 nodes. (0.20ms per tree) (0.000114ms per node)
refactored walker (with 2 switch)
total walking time 1.13sec for 4998 trees and 8770308 nodes. (0.23ms per tree) (0.000129ms per node)
heap based walker (with 2 switch)
total walking time 1.12sec for 4998 trees and 8770308 nodes. (0.22ms per tree) (0.000127ms per node)



run 3 *******
run gc to clean memory

total file size 69548977bytes, total parse time 15.65sec for 4998 trees (3.13ms per tree)
original walker
total walking time 0.96sec for 4998 trees and 8770308 nodes. (0.19ms per tree) (0.000110ms per node)
heap based walker (with 1 switch)
total walking time 0.99sec for 4998 trees and 8770308 nodes. (0.20ms per tree) (0.000113ms per node)
refactored walker (with 2 switch)
total walking time 1.13sec for 4998 trees and 8770308 nodes. (0.23ms per tree) (0.000129ms per node)
heap based walker (with 2 switch)
total walking time 1.1sec for 4998 trees and 8770308 nodes. (0.22ms per tree) (0.000125ms per node)



run 4 *******
run gc to clean memory

total file size 69548977bytes, total parse time 15.29sec for 4998 trees (3.06ms per tree)
original walker
total walking time 0.97sec for 4998 trees and 8770308 nodes. (0.19ms per tree) (0.000111ms per node)
heap based walker (with 1 switch)
total walking time 0.99sec for 4998 trees and 8770308 nodes. (0.20ms per tree) (0.000113ms per node)
refactored walker (with 2 switch)
total walking time 1.14sec for 4998 trees and 8770308 nodes. (0.23ms per tree) (0.000130ms per node)
heap based walker (with 2 switch)
total walking time 1.08sec for 4998 trees and 8770308 nodes. (0.22ms per tree) (0.000123ms per node)



run 5 *******
run gc to clean memory

total file size 69548977bytes, total parse time 15.62sec for 4998 trees (3.13ms per tree)
original walker
total walking time 0.97sec for 4998 trees and 8770308 nodes. (0.19ms per tree) (0.000111ms per node)
heap based walker (with 1 switch)
total walking time 0.98sec for 4998 trees and 8770308 nodes. (0.20ms per tree) (0.000112ms per node)
refactored walker (with 2 switch)
total walking time 1.12sec for 4998 trees and 8770308 nodes. (0.22ms per tree) (0.000128ms per node)
heap based walker (with 2 switch)
total walking time 1.1sec for 4998 trees and 8770308 nodes. (0.22ms per tree) (0.000125ms per node)
```

one idea to remove second switch is putting childnodes to each parse node itself rather than having extra function to get those. only problem is since parse node is interface, we need to put it as field which consume memory.